requirements.txt
python-telegram-bot==21.0  # Core library (no webhook extras)
python-dotenv==1.0.0       # For .env files (optional)
psutil==5.9.5              # Only if using process management
requests==2.31.0           # HTTP requests